refactored lidar and motor into their own files and started on the sensors facade class
